Refactor: Update `MainActivityUiState.Success` to use `Label` objects
This commit modifies the `MainActivityUiState.Success` data class.

The `labels` property, previously a `List<String>`, is now a `List<Label>`. This change aligns the UI state with the `Label` model object, providing richer data to the UI.

The `combine` function in `MainAppViewModel` has been updated to pass the `List<Label>` directly to the `Success` state, removing the previous mapping from `Label` objects to their `name` strings.